+++ src/storage/memory.ts
@@ -11,10 +11,15 @@ import { DocumentDriveStorage, DocumentStorage, IDriveStorage } from './types';
 export class MemoryStorage implements IDriveStorage {
     private documents: Record<string, Record<string, DocumentStorage>>;
     private drives: Record<string, DocumentDriveStorage>;
+    private slugToDriveId: Record<string, string> = {};
 
     constructor() {
         this.documents = {};
         this.drives = {};
+    }
+
+    checkDocumentExists(drive: string, id: string): Promise<boolean> {
+        return Promise.resolve(this.documents[drive]?.[id] !== undefined)
     }
 
     async getDocuments(drive: string) {
@@ -116,9 +121,20 @@ export class MemoryStorage implements IDriveStorage {
         return drive;
     }
 
+    async getDriveBySlug(slug: string) {
+        const driveId = this.slugToDriveId[slug];
+        if (!driveId) {
+            throw new Error(`Drive with slug ${slug} not found`);
+        }
+        return this.getDrive(driveId);
+    }
+
     async createDrive(id: string, drive: DocumentDriveStorage) {
         this.drives[id] = drive;
-        this.documents[id] = {};
+        const { slug } = drive.initialState.state.global;
+        if (slug) {
+            this.slugToDriveId[slug] = id;
+        }
     }
 
     async addDriveOperations(