+++ src/utils/index.ts
@@ -98,3 +98,4 @@ export function isNoopUpdate(
 export function isBefore(dateA: Date | string, dateB: Date | string) {
     return new Date(dateA) < new Date(dateB);
 }
+