+++ test/server.test.ts
@@ -5,7 +5,7 @@ import {
     reducer
 } from 'document-model-libs/document-drive';
 import * as DocumentModelsLibs from 'document-model-libs/document-models';
-import { DocumentModel } from 'document-model/document';
+import { ActionContext, DocumentModel } from 'document-model/document';
 import {
     actions as DocumentModelActions,
     DocumentModelDocument,
@@ -62,6 +62,8 @@ describe.each(storageLayers)(
 
             if (storageName === 'FilesystemStorage') {
                 return fs.rm(FileStorageDir, { recursive: true, force: true });
+            } else if (storageName === "BrowserStorage") {
+                return (await buildStorage()).clearStorage?.();
             } else if (storageName === 'PrismaStorage') {
                 await prismaClient.$executeRawUnsafe(
                     'DELETE FROM "Attachment";'
@@ -70,6 +72,9 @@ describe.each(storageLayers)(
                     'DELETE FROM "Operation";'
                 );
                 await prismaClient.$executeRawUnsafe('DELETE FROM "Document";');
+                await prismaClient.$executeRawUnsafe(
+                    'DELETE FROM "Drive";'
+                );
             }
         });
 
@@ -642,5 +647,110 @@ describe.each(storageLayers)(
                 document.operations
             );
         });
-    }
-);
+
+        it('saves operation context', async ({ expect }) => {
+            const server = new DocumentDriveServer(
+                documentModels,
+                await buildStorage()
+            );
+            await server.addDrive({
+                global: {
+                    id: '1',
+                    name: 'name',
+                    icon: 'icon',
+                    slug: 'slug'
+                },
+                local: {
+                    availableOffline: false,
+                    sharingType: 'public',
+                    listeners: [],
+                    triggers: []
+                }
+            });
+            let drive = await server.getDrive('1');
+
+
+            const context: ActionContext = {
+                signer: {
+                    user: {
+                        address: '123',
+                        networkId: '1',
+                        chainId: 1
+                    },
+                    app: {
+                        name: 'name',
+                        key: 'key'
+                    },
+                    signature: "test"
+                }
+            }
+
+            drive = reducer(
+                drive,
+                {
+                    ...DocumentDriveUtils.generateAddNodeAction(
+                        drive.state.global,
+                        {
+                            id: '1.1',
+                            name: 'document 1',
+                            documentType: 'powerhouse/document-model'
+                        },
+                        ['global', 'local'],
+
+                    ), context
+                }
+            );
+
+            // dispatches operation to server
+            const operation = drive.operations.global[0]!;
+            const operationResult = await server.addDriveOperation(
+                '1',
+                operation
+            );
+            expect(operationResult.status).toBe('SUCCESS');
+
+            drive = await server.getDrive('1');
+            expect(drive.operations.global[0]?.context).toStrictEqual(context);
+        });
+
+
+        it('get drives by slug', async ({ expect }) => {
+            const server = new DocumentDriveServer(
+                documentModels,
+                await buildStorage()
+            );
+            const addDrive = (driveId: string, slug: string) => server.addDrive({
+                global: {
+                    id: driveId,
+                    name: 'name',
+                    icon: 'icon',
+                    slug: slug
+                },
+                local: {
+                    availableOffline: false,
+                    sharingType: 'public',
+                    listeners: [],
+                    triggers: []
+                }
+            });
+
+
+            await addDrive('1', 'slug1');
+            await addDrive('2', 'slug2');
+            await addDrive('3', 'slug3');
+            await addDrive('4', 'slug1');
+
+
+            let drive = await server.getDriveBySlug('slug1');
+            expect(drive.state.global.id).toBe('4');
+
+            drive = await server.getDriveBySlug('slug2');
+            expect(drive.state.global.id).toBe('2');
+
+            drive = await server.getDriveBySlug('slug3');
+            expect(drive.state.global.id).toBe('3');
+        });
+
+    })
+
+